+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - Enh #67: Allow specifying delimiters for `*` (samdark)
 - Enh #67: Add `\Yiisoft\Strings\WildcardPattern::isDynamic()` (samdark)
 - Enh #67: Add `\Yiisoft\Strings\WildcardPattern::quote()` (samdark)
-- Chg #67: `\Yiisoft\Strings\WildcardPattern#withExactLeadingPeriod()` was removed (samdark)
+- Chg #67: Remove `\Yiisoft\Strings\WildcardPattern::withoutEscape()` (samdark)
+- Chg #67: Remove `\Yiisoft\Strings\WildcardPattern::withExactLeadingPeriod()` (samdark)
 
 ## 1.2.0 January 22, 2021